LPS-77454 Move web only methods into web internal CalendarResourceUtil.
